Back value on authentication dropdown dropdown transition value
In the checkout page when you close the login popup after opening you cannot click anywhere on the page. This is because the backdrop behind the modal doesn't get removed. The JavaScript for the modal at magento/module-ui/view/base/web/js/modal/modal.js looks at the CSS transition property to remove the backdrop. This value for the transition property is invalid that's why the transition property never get's picked up.