match core so 4 products appear per row
omg what a massive pain that was (and i prob should of gave up hours ago) trying to work out why rows incorrectly had three products, not four. I tried to cross ref the changes, but it was easier to convert the less to scss again. Turns out the calculations were outdated, regardless of how sass does or doesn't process the calc correctly.  

https://github.com/magento/magento2/commit/2b47de7e80cc7eaf3a97b036c0da2304914c2552#diff-f0f66ccc8eb7d2e0d4e821f28c6cf4663614761ecd6d47c2145dba0eb94ba5be